@@ -6,39 +6,34 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.15.0 <3.0.0"
+  sdk: ">=2.18.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-
   cupertino_icons: ^1.0.2
-
-  firebase_core: ^1.20.1
-  firebase_database: ^9.1.1
-  firebase_auth: ^3.6.3
-  firebase_storage: ^10.2.6
-  firebase_messaging: ^11.2.6
-  firebase_messaging_platform_interface: ^3.4.0
-
-
+  firebase_core: ^2.22.0
+  firebase_database: ^10.3.8
+  firebase_auth: ^4.14.0
+  firebase_storage: ^11.5.1
+  firebase_messaging: ^14.7.4
+  firebase_messaging_platform_interface: ^4.5.13
   get: ^4.6.5
-
   flutter_contacts: ^1.1.2
   flutter_multi_formatter: ^2.5.1
-  uuid: ^3.0.5
-  flutter_launcher_icons: ^0.10.0
-  image_picker: ^0.8.4+5
-  flutter_linkify: ^5.0.2
+  uuid: ^4.2.1
+  flutter_launcher_icons: ^0.13.1
+  image_picker: ^1.0.4
+  flutter_linkify: ^6.0.0
   cached_network_image: ^3.2.0
   bloc: ^8.0.2
   flutter_bloc: ^8.0.1
   path_provider: ^2.0.9
-  permission_handler: ^10.0.0
+  permission_handler: ^11.0.1
   open_file: ^3.2.1
   palestine_console: ^0.0.10
-  cloud_firestore: ^3.4.0
-  cloud_firestore_web: ^2.8.4
+  cloud_firestore: ^4.13.1
+  cloud_firestore_web: ^3.8.5
   get_storage: ^2.0.3
   emoji_picker_flutter: ^1.1.2
   video_player: ^2.4.5
@@ -48,7 +43,7 @@ dependencies:
   equatable: ^2.0.3
 
 dev_dependencies:
-  flutter_lints: ^2.0.1
+  flutter_lints: ^3.0.0
   msix: ^3.1.6
 
 flutter_icons:
@@ -61,16 +56,3 @@ flutter:
   uses-material-design: true
   assets:
     - assets/
-
-
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
