@@ -1,3 +1,3 @@
 import 'package:flutter/material.dart';
 
-TextStyle getUnavailableTextTheme(BuildContext context) => Theme.of(context).textTheme.subtitle2!.copyWith(color: Colors.grey);
+TextStyle getUnavailableTextTheme(BuildContext context) => Theme.of(context).textTheme.titleSmall!.copyWith(color: Colors.grey);
