@@ -4,15 +4,15 @@ import 'package:flutter/material.dart';
 
 class WelcomeToWhatsapp extends StatelessWidget {
   const WelcomeToWhatsapp({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
     return Center(
       child: Text(
         "BEM VINDO AO Clone do Zap",
-        style: Theme.of(context).textTheme.headline1,
+        style: Theme.of(context).textTheme.displayLarge,
         textAlign: TextAlign.center,
       ),
     );
