@@ -4,8 +4,8 @@ import 'package:flutter/material.dart';
 
 class FormHeader extends StatelessWidget {
   const FormHeader({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -14,7 +14,7 @@ class FormHeader extends StatelessWidget {
         Text(
           "Coloque seu número de telefone",
           textAlign: TextAlign.center,
-          style: Theme.of(context).textTheme.headline2,
+          style: Theme.of(context).textTheme.displayMedium,
         ),
         const Text(
           "O Clone do Zap vai precisar verificar o a sua identidade",
