@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 
 class ScaffoldLoading extends StatelessWidget {
   const ScaffoldLoading({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
