@@ -4,8 +4,8 @@ import 'package:flutter/material.dart';
 
 class AwesomeImage extends StatelessWidget {
   const AwesomeImage({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
