@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 
 class SearchInConversas extends StatelessWidget {
   const SearchInConversas({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
