@@ -1,14 +1,13 @@
 // ignore_for_file: invalid_use_of_protected_member, library_prefixes, constant_identifier_names, non_constant_identifier_names
 import 'package:flutter/material.dart';
 import 'package:whatsapp2/features/2_app/2_app_content/2.1_conversas/2_conversa/model/message_model.dart';
-import 'package:whatsapp2/features/2_app/2_app_content/2.1_conversas/2_conversa/widgets/3_conversas_texts/widgets/1_corpo/widgets/1_numero.dart';
 import 'package:whatsapp2/features/2_app/2_app_content/2.1_conversas/2_conversa/widgets/3_conversas_texts/widgets/1_corpo/widgets/2_foto.dart';
 import 'package:whatsapp2/features/2_app/2_app_content/2.1_conversas/2_conversa/widgets/3_conversas_texts/widgets/1_corpo/widgets/3_text.dart';
 import 'package:whatsapp2/features/2_app/2_app_content/2.1_conversas/2_conversa/widgets/3_conversas_texts/widgets/1_corpo/widgets/4_date.dart';
 
 class MensagemCorpo extends StatelessWidget {
   const MensagemCorpo({
-    Key? key,
+    super.key,
     required this.alignment,
     required this.MARGIN,
     required this.color,
@@ -20,7 +19,7 @@ class MensagemCorpo extends StatelessWidget {
     required this.dia,
     required this.constraints,
     required this.mensagem,
-  }) : super(key: key);
+  });
 
   final MainAxisAlignment alignment;
   final double MARGIN;
@@ -86,11 +85,11 @@ class MensagemCorpo extends StatelessWidget {
                           ),
                         )
                       : const SizedBox(),
-                  MensagemUserNumeroMaybe(
-                    nome: nome,
-                    numero: numero,
-                    showNumber: showNumber,
-                  ),
+                  // MensagemUserNumeroMaybe(
+                  //   nome: nome,
+                  //   numero: numero,
+                  //   showNumber: showNumber,
+                  // ),
                   MensagemArquivoMaybe(
                     myMensagem: myMensagem,
                     showNumber: showNumber,
