@@ -6,14 +6,15 @@ import 'package:get/get.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:whatsapp2/features/2_app/2_app_content/2.1_conversas/2_conversa/widgets/4_bottom_form/widgets/1_text_image_form/1_text_image_form.dart';
 import 'package:whatsapp2/features/2_app/3_configuracoes/widgets/1_user_identity/user_name.dart';
+
 import '../../../../../../common/widgets/eu/user_photo.dart';
 import '../../common/configuration_option.dart';
 import '../user_identity.dart';
 
 class ChangeUserDetails extends StatelessWidget {
   const ChangeUserDetails({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -91,8 +92,8 @@ class ChangeUserDetails extends StatelessWidget {
 
 class ChangePhotoBottomSheet extends StatelessWidget {
   const ChangePhotoBottomSheet({
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -101,8 +102,8 @@ class ChangePhotoBottomSheet extends StatelessWidget {
       child: Column(
         mainAxisSize: MainAxisSize.min,
         children: [
-          Row(
-            children: const [
+          const Row(
+            children: [
               Text(
                 "Foto do perfil",
                 style: TextStyle(
@@ -116,8 +117,8 @@ class ChangePhotoBottomSheet extends StatelessWidget {
             mainAxisSize: MainAxisSize.max,
             children: [
               TextButton(
-                child: Column(
-                  children: const [
+                child: const Column(
+                  children: [
                     Icon(Icons.photo_camera),
                     Text("Camera"),
                   ],
@@ -125,14 +126,15 @@ class ChangePhotoBottomSheet extends StatelessWidget {
                 onPressed: () {},
               ),
               TextButton(
-                child: Column(
-                  children: const [
+                child: const Column(
+                  children: [
                     Icon(Icons.image),
                     Text("Galeria"),
                   ],
                 ),
                 onPressed: () async {
                   Navigator.pop(context);
+                  print('pressed');
                   var user = FirebaseAuth.instance.currentUser;
                   XFile? image = await getImageFromGallery();
                   Get.showSnackbar(
@@ -142,13 +144,15 @@ class ChangePhotoBottomSheet extends StatelessWidget {
                       message: " ",
                     ),
                   );
-                  if (image == null || user == null) {
-                    return;
-                  }
+                  if (image == null) throw Exception('Image is null');
+                  if (user == null) throw Exception('User is null');
+
+                  print('image: ${image.path}');
                   var ref = FirebaseStorage.instance.ref('perfil/${user.uid}');
                   var result = await sendImageToFirebase(ref, image);
                   var imageUrl = await result.ref.getDownloadURL();
                   await user.updatePhotoURL(imageUrl);
+                  print('imageUrl: $imageUrl');
                   Get.showSnackbar(
                     const GetSnackBar(
                       duration: Duration(milliseconds: 2000),
